Implementing a form to submit quotes to the server through the client
Created an AddQuote component for Author, Quote and ImageURL fields and a button to submit

Updated actionCreators. to export the addQuote function with Author, Quote and imageURL paramters

updated sagas to import createQuote and add the function listener addQuoteSaga and the callback addQuote

updated API to export createQuote function

updated QuoteCard to include image and description.

However Images for quotes not being added when creating a quote, will need to look into it